pyspark-clean-architecture-example/requirements.txt
pyspark==4.0.1
delta-spark==4.0.1